Move server logic to its module
